# perf: cache shiki