--- a/src/node/contentLoader.ts
+++ b/src/node/contentLoader.ts
@@ -4,10 +4,7 @@ import { glob, type GlobOptions } from 'tinyglobby'
 import type { SiteConfig } from './config'
 import matter from 'gray-matter'
 import { normalizePath } from 'vite'
-import {
-  createMarkdownRenderer,
-  type MarkdownRenderer
-} from './markdown/markdown'
+import { createMarkdownRenderer } from './markdown/markdown'
 
 export interface ContentOptions<T = ContentData[]> {
   /**
@@ -100,15 +97,7 @@ export function createContentLoader<T = ContentData[]>(
   if (typeof pattern === 'string') pattern = [pattern]
   pattern = pattern.map((p) => normalizePath(path.join(config.srcDir, p)))
 
-  let md: MarkdownRenderer
-
-  const cache = new Map<
-    string,
-    {
-      data: any
-      timestamp: number
-    }
-  >()
+  const cache = new Map<string, { data: any; timestamp: number }>()
 
   return {
     watch: pattern,
@@ -124,14 +113,12 @@ export function createContentLoader<T = ContentData[]>(
         ).sort()
       }
 
-      md =
-        md ||
-        (await createMarkdownRenderer(
-          config.srcDir,
-          config.markdown,
-          config.site.base,
-          config.logger
-        ))
+      const md = await createMarkdownRenderer(
+        config.srcDir,
+        config.markdown,
+        config.site.base,
+        config.logger
+      )
 
       const raw: ContentData[] = []