### Description

<!-- Please insert your description here and provide info about the "what" this PR is solving. -->

Hello 👋,

I'm building my website, https://soubiran.dev, with VitePress.

At first, it was fast but as I added more content and more data loaders, it became slower and slower to build. Today, it takes around 4 minutes to build the website and each time I add a new page, it adds 10 seconds to the build time.

So I decided to investigate to find the annoying issue. At the same time, I had this warning: `Shiki is supposed to be used as a singleton, consider refactoring your code to cache your highlighter instance;`.

I moved all my content to a fresh VitePress project and I found that the speed build was back to normal. In my project, soubiran.dev, I started to play with `DEBUG=vitepress:md` and I discover that some pages took 27s to build while some was under 100ms. This led me to the conclusion that the issue was related to the markdown content and potentially the parser: MarkdownIt.

With the warning in mind, I quickly understand the highlighter creation was the issue: https://github.com/vuejs/vitepress/blob/main/src/node/markdown/markdown.ts#L208

I also discover that every time a `createContentLoader` is used, a new instance of MarkdownIt is created: https://github.com/vuejs/vitepress/blob/main/src/node/contentLoader.ts#L129

So I try to patch VitePress in my project by caching the highlighter instance.

Before, in the Cloudflare Pages CI: more than 4 minutes to build the website.
After, in the Cloudflare Pages CI: 47 seconds to build the website.

related to https://x.com/soubiran_/status/1851189188634288614

This also improve significantly the dev startup of the project.

**Unfortunately, this project is private but I can happily invite  maintainers if needed.**

### Linked Issues

<!-- e.g. fixes #123 -->

### Additional Context

<!-- Is there anything you would like the reviewers to focus on? -->

---

> [!TIP]
> The author of this PR can publish a _preview release_ by commenting `/publish` below.