@Barbapapazes 2024-10-29
_While I was writing this PR, I was wondering if we couldn't go further and cache the MardownIt instance?_