@brc-dd 2024-10-29
Actually the issue is we need to reload the highlighter instance and destroy the old one too in certain cases like reloading config file (because a user might change shiki's theme and expect it to reload). But yeah I'm seeing perf degradation with both cold start (and config reload).

If possible, can you add me temporarily with read-only access on the repo? You can dm me on discord if you've any concerns.

But from what you've told, I believe the issue in your case is because of using createMarkdownRenderer in data loader (similar to note in https://github.com/vuejs/vitepress/issues/4300#issuecomment-2439675326). We probably should provide better API instead of createMarkdownRenderer, or maybe an option to re-use the instance (enabled by default, but disabled inside our vite plugin so that on config reload it still creates and caches a new instance but inside data loaders, etc. it can re-use the cached one).