@Barbapapazes 2024-10-29
> /publish

I've updated soubiran.dev and this PR is in production! (build time: 30 seconds on CF, 13 seconds on my computer and a way better dev start time)