@Barbapapazes 2024-11-02
I just update my dependencies, and now, it's under 10 seconds 🥳

![Screenshot 2024-11-02 at 19 34 24](https://github.com/user-attachments/assets/a62f47c9-4568-41a6-b92f-246a3c22302f)